Cambios en formateo y se agregan y eliminan comentarios.

+++ src/main/java/ar/edu/unq/poo2/envio/EnvioEstandar.java
@@ -16,7 +16,7 @@ public class EnvioEstandar implements MetodoDeEnvio {
 		double pesoTotal = pedido.getPesoTotal(); 
         Direccion destino = pedido.getDireccionEntrega();
         
-        return (double) correo.estimarEnvio(pesoTotal, destino);
+        return correo.estimarEnvio(pesoTotal, destino);
 	}
 
 	@Override
